+++ HISTORY.rst
@@ -6,6 +6,8 @@ History
 0.2.0 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 
+- Allow custom serialization in JSONRenderer without abusing __repr__.
+
 
 0.1.0 (2013-09-12)
 ^^^^^^^^^^^^^^^^^^